requirements.txt
pygame>=2.0.0
numpy>=1.20.0
watchdog>=2.1.0
rich>=14.1.0
Pillow>=9.0.0
psutil>=5.8.0
pyaudio>=0.2.11
